rbsync/requirements.txt
mysql-connector-python
python-dotenv
pdftotext
